requirements.txt
semantic-kernel
python-dotenv
chainlit